ARM: tegra14: emc: Add emc_timing_update
Add an emc_timing_update() that had been previously removed.
This update is necessary even though the register being written
to just before hand is not shadowed.

The emc_timing_update() inserts a stall into the EMC pipeline.
The autocal SM waits for a stall before running the autocal.
Without this stall, the EMC can wait too long for an autocal
and end up getting invalid reads which crash the system.

Change-Id: Ib28047e23f01bdeba634d0bb225f30aef9cf9ef3
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/333675
Signed-off-by: Rick Song <ricks@nvidia.com>
Reviewed-on: http://git-master/r/366330